Added more 0x140c1df14 skips
I froze once after playing for several hours, it was in the Open Zone - but the "cyberspace" code runs in both Cyberspace and Open Zone, affecting Cyberspace the most.
It is possible the freeze I experienced was some new code we have not found yet - I was not running my profiler.
There are five other paths I know of that reach 0x140c1df14, that being:
- 0x14EA6D7C1
- 0x14EA6D7D5
- 0x14EA6D3B8
- 0x14EA6CD64
- 0x14EA6CDD4
But I experienced immediate issues trying to skip these. These might be worth reinvestigating in the future though as I only tried each once or twice and I might have made a mistake.